requirement.txt
Babel==2.3.4
Django==1.10.3
mysqlclient==1.3.9
oslo.i18n==3.10.0
requests==2.12.1
django-settings-export==1.2.1
pytz==2016.10
docker==2.2.0
dateutils==0.6.6
django-cuser==2017.3.16